requirements.txt
# AI Employee System - Python Dependencies

# Google Gmail API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Filesystem monitoring
watchdog>=3.0.0

# Environment variable management
python-dotenv>=1.0.0
